style: run prettier

+++ src/modules/like/like.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Param, Body, Post } from '@nestjs/common';
+import { Controller, Param, Body, Post, HttpException } from '@nestjs/common';
 import { LikeService } from './like.service';
 import { LikeRequest } from './dtos/like.request';
 import { LikeDto } from './dtos/like.dto';
@@ -7,15 +7,18 @@ import { LikeDto } from './dtos/like.dto';
 export class LikeController {
     constructor(private likeService: LikeService) {}
 
-    @Post('/:userId')
-    async addLike(@Param('userId') userId: string, @Body() likeReq: LikeRequest): Promise<LikeDto> {
-        const like = await this.likeService.addLike(userId, likeReq.photoId);
+    @Post('/addLike')
+    async addLike(@Body() likeReq: LikeRequest): Promise<LikeDto> {
+        const like = await this.likeService.addLike(likeReq.userId, likeReq.photoId);
         return LikeDto.from(like);
     }
 
-    @Post('/:userId')
-    async removeLike(@Param('userId') userId: string, @Body() likeReq: LikeRequest): Promise<LikeDto> {
-        const like = await this.likeService.getLikeByPhotoId(userId, likeReq.photoId);
+    @Post('/removeLike')
+    async removeLike(@Body() likeReq: LikeRequest): Promise<LikeDto> {
+        const like = await this.likeService.getLikeByPhotoId(likeReq.userId, likeReq.photoId);
+        if (!like) {
+            throw new HttpException(`No like on this photo`, 403);
+        }
         await this.likeService.removeLike(like.id);
         return LikeDto.from(like);
     }
